requirements.txt
appnope==0.1.0
attrs==19.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
configparser==3.7.4
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
entrypoints==0.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
GDAL==2.4.0
glumpy==1.1.0
greenlet==0.4.15
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
multiprocess==0.70.7
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
numpy-stl==2.10.1
pandocfilters==1.4.2
pathlib2==2.3.3
pathos==0.2.3
pexpect==4.7.0
pickleshare==0.7.5
pox==0.2.5
ppft==1.6.4.9
progressbar2==3.39.3
prometheus-client==0.6.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.3.1
pymesh==1.0.2
PyOpenGL==3.1.0
pyparsing==2.4.0
pyroSAR==0.8
pyrsistent==0.15.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
scandir==1.10.0
scoop==0.7.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
spatialist==0.2.8
subprocess32==3.5.3
tblib==1.3.2
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
triangle==20190115.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
